ファイルの移動・整理 DiaryList class
