+++ lib/diary_detail/diary_detail.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'diary.dart';
+import '../diary.dart';
 
 class DiaryDetail extends StatelessWidget {
   final Diary diary;
